Using WalletView on ExecutionDialog.
The WalletView nedded to be changed to also look for id, but I should
have done that way from beggining anyway, so it was good.

The WalletOverviewDialog had to be changed accordingly to new interface.
But it still depends on positional, the id parameter is ignored for now
(a issue was open to fix it).