Basic Undo/Redo support working (Issue #12);
Everything is working at model level, but only the estimates grid will respond well.
There is no way to restrict the undo/redo to a dialog scope nor to notify the budgetList that
budget changed. The forme can go unoticed, as the undo/redo doesn't work on these dialogs, but
the later need to be addressed.